+++ LookupAnything/release-notes.md
@@ -4,7 +4,8 @@
 ## Upcoming release
 * Updated for Stardew Valley 1.5, including...
   * split-screen mode and UI scaling;
-  * added island crystal cave puzzle lookup.
+  * added island crystal cave puzzle lookup;
+  * added island shrine puzzle lookup.
 * Fixed some untranslated text.
 * Fixed NPC lookups not showing social info for some custom mod NPCs. (This affects NPCs with friendship data but `CanSocialize` set to false.)
 * Improved translations. Thanks to spindensity (updated Chinese)!